requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
cffi==1.15.0
colorama==0.4.4
cssselect2==0.5.0
dj-database-url==0.5.0
Django==4.0
django-cors-headers==3.10.1
django-countries==7.2.1
django-easy-pdf==0.1.1
django-heroku==0.3.1
django-pandas==0.6.6
django-phone-field==1.8.1
django-phonenumber-field==6.0.0
djangorestframework==3.13.0
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
lxml==4.8.0
milksnake==0.1.5
numpy==1.22.2
pandas==1.4.1
Pillow==9.0.1
psycopg2==2.9.3
pycparser==2.21
PyPDF2==1.26.0
PyPDF3==1.0.6
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2021.3
reportlab==3.6.6
six==1.16.0
sqlparse==0.4.2
svglib==1.2.1
tinycss2==1.1.1
tqdm==4.63.1
tzdata==2021.5
urlquote==1.1.4
webencodings==0.5.1
whitenoise==5.3.0
xhtml2pdf==0.2.5
